backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp==3.6.2            # via alpha-vantage
alembic==1.4.2            # via -r requirements.in, flask-migrate
alpha-vantage==2.2.0      # via -r requirements.in
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp
babel==2.8.0              # via flask-babelex
bcrypt==3.1.7             # via -r requirements.in, flask-user
blinker==1.4              # via flask-mail
certifi==2020.6.20        # via requests
cffi==1.14.0              # via bcrypt, cryptography
chardet==3.0.4            # via aiohttp, requests
click==7.1.2              # via -r requirements.in, flask
cryptography==2.9.2       # via -r requirements.in, flask-user, pyopenssl
dnspython==1.16.0         # via email-validator
email-validator==1.1.1    # via -r requirements.in
flask-babelex==0.9.4      # via -r requirements.in
flask-cors==3.0.8         # via -r requirements.in
flask-login==0.5.0        # via flask-user
flask-mail==0.9.1         # via flask-user
flask-migrate==2.5.3      # via -r requirements.in
flask-script==2.0.6       # via -r requirements.in
flask-sqlalchemy==2.4.3   # via -r requirements.in, flask-migrate, flask-user
flask-user==1.0.2.2       # via -r requirements.in
flask-wtf==0.14.3         # via flask-user
flask==1.1.2              # via -r requirements.in, flask-babelex, flask-cors, flask-login, flask-mail, flask-migrate, flask-script, flask-sqlalchemy, flask-user, flask-wtf
gunicorn==20.0.4          # via -r requirements.in
idna==2.9                 # via email-validator, requests, yarl
inflection==0.5.0         # via quandl
itsdangerous==1.1.0       # via flask, flask-wtf
jinja2==2.11.2            # via flask, flask-babelex
joblib==0.16.0            # via scikit-learn
lxml==4.5.1               # via pandas-datareader
mako==1.1.3               # via alembic
markupsafe==1.1.1         # via jinja2, mako, wtforms
marshmallow==3.7.1        # via -r requirements.in, webargs
more-itertools==8.4.0     # via quandl
multidict==4.7.6          # via aiohttp, yarl
multitasking==0.0.9       # via yfinance
numpy==1.19.0             # via -r requirements.in, pandas, quandl, scikit-learn, scipy, yfinance
pandas-datareader==0.8.1  # via -r requirements.in
pandas==1.0.5             # via -r requirements.in, pandas-datareader, quandl, yfinance
passlib==1.7.2            # via -r requirements.in, flask-user
psycopg2==2.8.5           # via -r requirements.in
pycparser==2.20           # via cffi
pyopenssl==19.1.0         # via -r requirements.in
python-dateutil==2.8.1    # via alembic, pandas, quandl
python-dotenv==0.13.0     # via -r requirements.in
python-editor==1.0.4      # via alembic
pytz==2020.1              # via babel, pandas
quandl==3.5.1             # via -r requirements.in
redis==3.5.3              # via -r requirements.in
requests==2.24.0          # via alpha-vantage, pandas-datareader, quandl, yfinance
safe==0.4                 # via -r requirements.in
scikit-learn==0.23.1      # via sklearn
scipy==1.5.1              # via scikit-learn
six==1.15.0               # via bcrypt, cryptography, flask-cors, pyopenssl, python-dateutil, quandl, sqlalchemy-utils
sklearn==0.0              # via -r requirements.in
speaklater==1.3           # via flask-babelex
sqlalchemy-utils==0.36.6  # via -r requirements.in
sqlalchemy==1.3.17        # via alembic, flask-sqlalchemy, sqlalchemy-utils
threadpoolctl==2.1.0      # via scikit-learn
urllib3==1.25.9           # via requests
validate-email==1.3       # via -r requirements.in
webargs==6.1.0            # via -r requirements.in
werkzeug==1.0.1           # via flask
wtforms==2.3.1            # via flask-wtf
yarl==1.4.2               # via aiohttp
yfinance==0.1.54          # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
